fire page scroll after tokens load
